% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{cdm_disconnect}
\alias{cdm_disconnect}
\title{Disconnect from a CDM database}
\usage{
cdm_disconnect(con)
}
\arguments{
\item{con}{A live \code{DBI} connection returned by \code{\link[=cdm_connect]{cdm_connect()}}.}
}
\value{
Invisibly returns the result of \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}}.
}
\description{
Disconnect from a CDM database
}
